requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
matplotlib>=3.5.0
flask>=2.3.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Security
bandit>=1.7.0
safety>=2.3.0
